Move order creation into order sub component
