+++ website/src/components/OrderReview.vue
@@ -1,0 +1,15 @@
+<template>
+  <v-container>
+    <h1> Review Order </h1>
+
+  </v-container>
+</template>
+<script>
+export default {
+  name: 'OrderReview',
+  store,
+}
+
+</script>
+<style>
+</style>
